moved files around to be set up by features
