+++ __tests__/Home.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Home from '../src/scenes/Home/Home';
+import Home from '../src/Home/Home';
 import renderer from 'react-test-renderer';
 import toJson from 'enzyme-to-json';
 import { shallow } from 'enzyme';
