+++ __tests__/NavBar.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import NavBar from '../src/components/NavBar/NavBar';
+import NavBar from '../src/NavBar/NavBar';
 import renderer from 'react-test-renderer';
 import toJson from 'enzyme-to-json';
 import { shallow } from 'enzyme';
