+++ src/Register/Register.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Auth } from '../../firebase/auth';
+import { Auth } from '../firebase/auth';
 import './Register.css';
 
 export default class Register extends Component {
